fix(react-doctor): move deriveRotation out of the PassportStamp component file
react-doctor's only-export-components check flagged the exported
deriveRotation helper: a component file that also exports a
non-component value breaks Fast Refresh's ability to preserve state.
Moved it to a sibling module (passport-stamp-rotation.ts) so
PassportStamp.tsx goes back to exporting only the component.
